Add files via upload
This Python script utilizes the K-Nearest Neighbors (KNN) algorithm to classify patients with diabetes based on their medical data. It begins by loading the diabetes dataset and conducting data cleaning and visualization for better insight. Standard scaling is then applied to normalize the data, enhancing the model's performance. The dataset is split into training and testing sets, and the KNN model is trained with the former. Performance evaluation metrics such as accuracy, precision, recall, and F1-score are computed to assess the model's effectiveness. Additionally, the relationship between the number of neighbors and model accuracy is visualized to identify the optimal neighbor count. A confusion matrix is employed for further performance visualization. Ultimately, the script concludes with a summary indicating that the KNN model demonstrates a commendable accuracy in diabetes patient classification.